LS-2169 - on paper, it's complete. time to test

+++ listshop/src/commonMain/kotlin/com/listshop/bff/data/remote/ApiShoppingList.kt
@@ -54,7 +54,7 @@ data class ApiShoppingListItem(
     @SerialName("tag_name")
     val tagName: String?,
     @SerialName("used_count")
-    val usedCount: Long?,
+    val usedCount: Int?,
     @SerialName("source_keys")
     val sourceKeys: List<String>?,
     @SerialName("tag")
